Conflicts:
	doc/modules/linear_model.rst

+++ sklearn/linear_model/base.py
@@ -48,10 +48,12 @@ def sparse_center_data(X, y, fit_intercept, normalize=False):
     if fit_intercept:
         # we might require not to change the csr matrix sometimes
         # store a copy if normalize is True.
+        # Change dtype to float64 since mean_variance_axis0 accepts
+        # it that way.
         if sp.isspmatrix(X) and X.getformat() == 'csr':
-            X = sp.csr_matrix(X, copy=normalize)
+            X = sp.csr_matrix(X, copy=normalize, dtype=np.float64)
         else:
-            X = sp.csc_matrix(X, copy=normalize)
+            X = sp.csc_matrix(X, copy=normalize, dtype=np.float64)
 
         X_mean, X_var = mean_variance_axis0(X)
         if normalize:
